# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger для АС ИДК.",
    "version": "0.1.0",
    "description": " ФЯО ФГУП ГХК, ДИТ 2023 г. ",
    "license": {
      "name": "ДИТ",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Ссылка на гит",
      "url": "https://github.com/zhiraff/idc-back"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/references/user/": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Показать всех пользователей",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "description": "Созданпе новой записи",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "secondName": {
                  "example": "any"
                },
                "thirdName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/references/user/search": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Поиск пользователей",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "query",
            "type": "string"
          },
          {
            "name": "secondName",
            "in": "query",
            "type": "string"
          },
          {
            "name": "thirdName",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/references/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Показать пользователя подробно",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "user"
        ],
        "description": "Обновление записи",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "secondName": {
                  "example": "any"
                },
                "thirdName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "description": "Удаление записи",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}